hw 8: allocate TimeStampDecoratorImpl and MessageGeneratorImpl
