+++ src/main/java/com/tcs/edu/interfaces/MessageGenerator.java
@@ -1,0 +1,7 @@
+package com.tcs.edu.interfaces;
+
+import com.tcs.edu.domain.*;
+
+public interface MessageGenerator {
+    void generateMessage(Message[] messageWithoutDoubles, Message[] messages);
+}
