+++ src/main/java/com/tcs/edu/interfaces/TimeStampDecorator.java
@@ -1,0 +1,7 @@
+package com.tcs.edu.interfaces;
+
+import com.tcs.edu.domain.*;
+
+public interface TimeStampDecorator {
+    String addTimeStamp();
+}
